requirements.txt
aioredis==1.3.1
appdirs==1.4.4
ariadne==0.12.0
async-timeout==3.0.1
cffi==1.14.3
cfgv==3.2.0
click==7.1.2
cryptography==3.2
distlib==0.3.1
filelock==3.0.12
graphql-core==3.0.5
h11==0.10.0
hiredis==1.1.0
identify==1.5.5
jwt==1.0.0
nodeenv==1.5.0
passlib==1.7.2
pre-commit==2.7.1
pycparser==2.20
pydantic==1.6.1
PyJWT==1.7.1
PyYAML==5.3.1
six==1.15.0
SQLAlchemy==1.3.19
starlette==0.13.8
toml==0.10.1
typing-extensions==3.7.4.3
uvicorn==0.12.1
virtualenv==20.0.33
